LibGfx+JPGLoader: Add support for chroma subsampling
This patch adds support for chroma subsampling in the JPGLoader.
We support subsampling standards defined by JFIF, which is to sample
half of chroma horizontally, half of chroma vertically, and chroma
quartered in both horizontal and vertical directions. Now that we can
handle subsampled chroma channels, the browser should be able to load
most of JPEGs (JFIF) found on the web, provided that images strictly
comply to the JFIF standard.