LibGfx+JPGLoader: Fix a bug in the decoder and prevent browser crash
When chroma subsampling is used, the decoder didn't resize the MCU
vector to account for the padded MCUs. The reason for this was that
the padded MCU vector length was calculated before reading subsampling
factors, which always resulted in the "padded size" to be equal to the
size of the MCU vector without padding. This patch fixes that.

In addition to that, this patch also modifies `get_next_symbol` to not
crash the browser with an assert when the generated codes don't match
the codes in huffman tables. Since this will often happen when we come
across JPEGs that don't follow JFIF standards or use entropy encoding
techniques other than huffman encoding, it's reasonable to just give
up on those images and move on.