AK: Support boundless peeking with BufferStream
This is an API enhancement. Boundless peeking is helpful when we have
to decide something based on the input before reading ahead. Here is an
example:

    bool id3_tag_ahead(BufferStream& stream)
    {
        auto data = stream.peek(3);
        return data[0] == 'I' && data[1] == 'D' && data[2] == '3';
    }

    while (id3_tag_ahead(stream) {
        parse_id3_tag(stream);
    }

    parse_mp3_frame(stream);

Without boundless peeking,if we have parsed all the ID3 tags then we'll
end up consuming 3 bytes of frame data. To counter that, we'll have to
pass the 3 bytes to the `parse_mp3_frame` which is not cool. Judging by
the API, current IO implementations seem to read entire files at once (
correct me if I am wrong here). This makes peeking even easier since we
don't have to switch between multiple buffers when boundaries are met.