AK: Add BinaryStream for binary parsing
This is a bare-bone implementation of an input stream specifically for
parsing binary files. With this, it becomes less of a pain to parse
compressed binary files having multibyte data entities in big-endian
format. `BinaryStream` is specially helpful when the format requires
reading beyond byte boundaries. The implementation is not complete, as
we ignore the possibility that the host machine itself can be a big
endian device. Furthermore, reading beyond byte boundaries is only
supported while reading big endian data. Manual shifting is required
if the bits need to be in little endian format.

Right now, it looks like this class doesn't add anything other than the
`read_network_order_bits` function to `BufferStream`. But I (or someone
else in need) will add more and more functions that prove helpful while
parsing binary files.