LibAudio: Start implementing ID3v2 Parser
Initial implementation of a basic ID3 parser. ID3 tags provide metadata
to MP3 files. In practice, we can completely ignore the ID3 tags and
seek to the first frame of the audio stream, but metadata contains
useful things like album art and genre. We will not try to parse every
tag the the ID3 standard defines, but a subset that we are interested
in.