MpegAudioLoader: strip comments and restructure parts of the decoder
This restructuring makes the implementation slightly disconnected from
the standard, but simplifies the design a lot. In other words, if one
was to make sense of the decoder by reading the spec word-by-word, they
will have a hard time. But someone who understands the MP3 format should
have no problem reasoning about the slight deviation of structs from
their definition in the spec.